Added timestamp & expiration

+++ bot.js
@@ -6,30 +6,39 @@ const intents = new Discord.Intents(32767);
 
 const client = new Discord.Client({ intents });
 
-const channelId = "919984229176201269";
+const channelId = "920332564831563807";
 
+var createdAt;
 var currentPollId;
 
-const createPoll = async () => {
+const createPoll = async (pingPlayers) => {
 	const channel = await client.channels.cache.get(channelId);
-	const message = await client.channels.cache.get(channelId).send("Ajoutez une réaction pour vous déclarer comme volontaire pour un 5v5 !");
+	channel.bulkDelete(99);
+	createdAt = new Date();
+	const message = await channel.send(`Ajoutez une réaction pour vous déclarer comme volontaire pour un 5v5 !
+Quand 10 personnes se seront portées volontaires, vous serez ping pour venir jouer (reset toutes les 4h).
+${pingPlayers ? "\n <@&823636943736275005>" : ""}
+\n(Ecrire "new" dans ce channel pour lancer un nouveau rassemblement)
+\n \n \n Création : ${createdAt.toLocaleString()}`);
 	await message.react('✅');
-	return message.id;
+	currentPollId = message.id;
 }
 
 client.once("ready",async () => {
 	console.log("5v5 bot online");
-	const channel = await client.channels.cache.get(channelId);
-	channel.bulkDelete(99);
-	currentPollId = await createPoll();
-	console.log(`CurrentPollId : ${currentPollId}`);
+	await createPoll(false);
+	// console.log(`CurrentPollId : ${currentPollId}`);
 });
 
 client.on('messageReactionAdd', async (reaction,user) =>{
 	if (user.bot || reaction.message.channelId !== channelId || reaction.message.id !== currentPollId) return;
+	let now = new Date();
+	if (now - createdAt > 4*60*60*1000){
+		createPoll(false);
+	}
 	const count = reaction.count - 1;
 	console.log(`Count : ${count}`);
-	if (count > 0){
+	if (count == 10){
 		reaction.users.fetch()
 			.then((users) => {
 				let msg = '';
@@ -42,9 +51,14 @@ client.on('messageReactionAdd', async (reaction,user) =>{
 	}
 });
 
-client.on("messageCreate", (message) => {
+client.on("messageCreate", async (message) => {
 	if (message.author.bot || message.channelId !== channelId) return;
-	message.delete();
+	if (message.content === "new"){
+		await createPoll(true);
+	}
+	else{
+		message.delete();
+	}
 });
